ip_grib2 branch:
Modify gdswzd routine and its regression test
to use a grid definition template array 
(variable igdtmpl) that is four bytes.  This
is how the ncep g2 library defines igdtmpl.

Keep two versions of the gdswiz_wzd regression
test source code: one for the branch library and 
one for the trunk.  This is because the gdswzd interface
has been changed under this ticket.  Link to
the correct source version in the make_reg_tests.ksh
build script.

Ticket #81.