ip_grib2 branch:
Routine gdswzd00, which does the grid transforms for
lat/lon grids, was modified to set the earth radius
from routine "earth_radius" instead of using a
parameter statement.  Add error check to ensure
the correct gdt template number is selected and
the earth radius is properly specified.  If either not
correct, set output fields to fill values and
exit.  Modify the gdswiz_wzd regression test for
grid 3, which is a global lat/lon grid.  The radius
of the earth (section 3, octet 15) was set wrong.
The correct value for that octet is '6', a 
spherical earth of 6371229 meters.

Ticket #81.