ip_grib2 test:
Fix error in gdswiz_wzd regression test.  The 
specification of the resolution and component
flags for grids 3 and 8 were changed from '24'
to '56' (bit 3 should be '1', not '0').  The
iplib does not use this bit, so results were
unaffected.

Baseline an initial version of a copygb2 
regression test.  Test will interpolate
a grib2 dataset of max snow albedo or 
elevation on a lat/lon grid to several
ncep and afwa standard grids.  This test
will be a replacement for the copygb
regression test, which works off grib 1.

See ticket #81 for more details.