ip/c_interop branch:
Baseline a unit test for testing the 'c' wrapper.

The program source code is in ./sorc.  It is 
based on Dusan's program. 

Links to the iplib library and include files
required by the program are in ./lib.

The executables will reside in ./exec.

Scripts to run the program are in ./scripts

This initial version is only for wcoss
phase 1/2.

Ticket #97.