trunk: ip
Per NCO requirements, add the "BSUB -cwd" option to all WCOSS 
unit test scripts to specify the current working directory. 
Prior to this requirement, the default cwd was the directory 
from which the script was submitted to lsf. Hence, I set the 
option as "-cwd $(pwd)" to preserve the current functionality.

Also, the scripts were modified to invoke lsf directly because 
variables such as $pwd are not recognized by the BSUB command 
when a script is invoked indirectly. As a result, I changed 
the script file extensions from ".lsf" to ".ksh".

Scripts affected are:

 reg_tests/Runall.wcoss-cray.ksh
 reg_tests/Runall.wcoss.ksh
 unit_test/scripts/run_wcoss-cray.ksh (renamed from .lsf)
 unit_test/scripts/run_wcoss.ksh (renamed from .lsf)
 unit_test_for_c/scripts/run_wcoss-cray.ksh (renamed from .lsf)
 unit_test_for_c/scripts/run_wcoss.ksh (renamed from .lsf) 

All associated README files were updated accordingly.

Since the changes were trivial, they were not documented
under a ticket.