Combine ip and ip2 libraries (#54)
ip and ip2 were combined using modern Fortran to abstract away the grib and grib2 specific portions into a unified interface. 

The API is backwards compatible, but some changes may be necessary to satisfy type checking now that there are module interfaces.

All routines were put into modules for type checking and other benefits.

The entry point into the library is `use ip_mod` which contains the public API (ipolates, ipolatev, gdswzd).

Doxygen documentation was started.

The raw grib1/grib2 grid descriptor is used to create an abstract ip_grid_descriptor which generalizes accessing common descriptor information reducing the need for grib1/grib2 specific code outside of descriptor initialization.

Then, that descriptor is used to create an ip_grid object which contains the grid-specific transformations.

Now, the actual interpolation routines work on ip_grids without having to know whether it was created by a grib1 or grib2 descriptor. The majority of the code can be shared. There are times when grib1/2 specific code is needed, and that can be achieved with a Fortran `select type` on the abstract type to access the concrete grid/descriptor type.

The code was tested with the combined unit tests of ip and ip2, UFS_UTILS' regressions tests, and wgrib2.